refactor(surface-remote): one evidence-line schema, not two plus a blind guard
A3 (reuse reviewer, verified with tsc). `FailureEvidenceSchema` in
surface-map's `define.ts` and `LogEntrySchema` in surface-remote's
`connection.ts` were character-for-character the same zod object. The previous
round made `LogEntry = EvidenceLine` (one TYPE) but left both zod definitions
standing, guarded by a `ZodType<LogEntry>` annotation documented as the drift
guard. It is not one: TypeScript accepts a NARROWER schema annotated as a wider
type, so adding a third `source` provenance to `EvidenceLine` would have left
`LogEntrySchema` compiling clean while rejecting the new value at runtime.

The stated reason for keeping the schema module-private — "exporting it would
invite a SECOND validation site" — was already false when written:
`connection.ts` validates these same lines inside `ConnectionInfoSchema.log`
and always has. Privacy bought a second DEFINITION of a site that existed
anyway.

BUNDLE CHECK (the finding asked for it): a value import from `@kolu/surface-map`
WOULD have dragged the contract builder into the browser graph — the default
entry re-exports `./define`, whose VALUE imports are `@orpc/contract`,
`@kolu/surface/define`, `zod` and `./envelope`. So rather than break
`connection.ts`'s documented "zod + type-only shapes" constraint, the
vocabulary moves to a new zod-only leaf, `@kolu/surface-map/evidence`
(`EvidenceLine`, `EvidenceLineSchema`, `FailureEvidence`,
`FailureEvidenceSchema`), re-exported from `./define` and `.` so nothing else
changes its imports. `LogEntrySchema` is now `EvidenceLineSchema` — an alias
has no direction to be blind in, so no replacement guard is needed.

Reference pages updated for the new subpath and exports (surface-reference
rule). No behaviour change; drishti needs no adoption — every previously
exported name still resolves from the same paths.
