+++ packages/surface-map/src/index.ts
@@ -14,7 +14,9 @@ export {
   type EntryStatus,
   entryStatusSchema,
   type EvidenceLine,
+  EvidenceLineSchema,
   type FailureEvidence,
+  FailureEvidenceSchema,
   type FailureRecord,
   type Key,
   type KeyCodec,
