+++ website/src/content/surface/ref-surface-map.mdx
@@ -11,7 +11,10 @@ import Snippet from "../../components/docs/Snippet.astro";
 A dynamic keyed map of remote surfaces — one entry surface, typed once, keyed at
 runtime, served as one. It depends only on `@kolu/surface` and has three public
 subpaths: the contract, the server, and the client (plus a test-only `/testing`
-subpath, whose `testMembershipId()` mints a branded `MembershipId` for fixtures).
+subpath, whose `testMembershipId()` mints a branded `MembershipId` for fixtures, and a
+zod-only `/evidence` leaf carrying the failure-evidence vocabulary — `EvidenceLine`,
+`EvidenceLineSchema`, `FailureEvidence`, `FailureEvidenceSchema`, all also re-exported
+from `.` — for a consumer that must not pull the contract builder into its bundle).
 
 ## The three halves
 
@@ -36,9 +39,12 @@ string-keyed. The domain failure schema is the `failure` argument itself — emb
 `entriesSpec.schema` and the `entries.get` contract output, so there is no separate
 `failureSchema` field. There is deliberately **no** matching `evidence` option: failure
 evidence is a fixed structural type this package owns (`FailureEvidence` and
-`EvidenceLine`, both exported from `.`), not a domain value to parameterize — see
-`EntryStatus` below. Its zod schema is module-private: evidence reaches the wire through
-the one door `entryStatusSchema` owns, so there is no second validation site to hold. The optional **`entriesClient`** (SR11) is the app-typed
+`EvidenceLine`, both exported from `.` and from `/evidence`), not a domain value to
+parameterize — see `EntryStatus` below. Their zod schemas (`FailureEvidenceSchema`,
+`EvidenceLineSchema`) are exported too: evidence reaches the *map's* wire through the one
+door `entryStatusSchema` owns, but the same line vocabulary is validated by
+`@kolu/surface-remote`'s `ConnectionInfoSchema.log`, which now *is* `EvidenceLineSchema`
+rather than a re-declared twin. The optional **`entriesClient`** (SR11) is the app-typed
 client error policy (`ClientCollectionPolicy<EP>`) threaded onto `entriesSpec.client`,
 so a map-membership subscription failure reaches the app's `connectSurfaceMap`
 `onClientError`; its `EP` is inferred from the value and defaults to `never` (a
@@ -214,7 +220,7 @@ watchByEntry(client, cell, items, onRaise) → { get(key): { kind: "live" | "sta
 - **`Conn` is the fine connection payload (SR9), carried on the `warming` and `connected` arms.** Optional and inferred from the map's `connection` schema, exactly like `Failure`: the package carries `state.connection` and validates it against that schema, never enumerating what a domain's connection states are. It is the ONE authority the coarse dot (`kind`) and the fine word both derive from — `serveHostMap` produces both from the same `SessionState` frame in one projection, so a "dot connected, word connecting" split (the drishti#102 divergence) has no construction path. A consumer reads the word off the same entry it reads the dot from (`state().connection`), never a second subscription; a map with no `connection` schema simply omits the field. The **`failed` arm carries no `connection` at all** — see the evidence bullets below for why that field's absence is the point rather than an oversight.
 - **The `failed` arm carries a schema-valid domain `failure` value**, validated by the map's own `failure` schema — not a loose object. padi's `PadiEntryFailure` is a `z.discriminatedUnion` over a structural `cause`, each arm carrying its human `reason` (and the `contract-skew-refused` arm additionally a typed `running`/`expected` version pair), so a consumer reads `state.failure.cause` / `state.failure.reason` (and `.running`/`.expected` on the skew arm) with full narrowing.
 - **The `failed` arm also carries the failure's `evidence`**, and it is **required**: `FailureEvidence = readonly { source: "local" | "remote"; line: string }[]`, the retained output tail of the episode that failed. A failed status without it fails `entryStatusSchema` — enforcement at the codec, so "a reason without its evidence" cannot even be decoded off the wire. `EvidenceLine.source` says **where** the line came from — `"local"` is the serving process's own output (provisioning, lifecycle chatter), `"remote"` is the far end's forwarded output — and it is a **field**, never an in-band `[local] ` prefix a reader would have to parse back off the string. The two together are `FailureRecord<Failure>`, exported from `.`.
-- **Evidence is a fixed structural type the package owns, not a third generic.** Unlike `Failure` and `Conn`, evidence is not domain volatility: any transport that can fail can produce provenance-tagged output lines, and `@kolu/surface-remote`'s `SessionState.log` is structurally exactly this — so the producer passes its retained tail straight through, with no injection and no second evidence pipe. There is only one vocabulary to keep straight: `@kolu/surface-remote`'s `LogEntry` **is** `EvidenceLine` (a type alias, with its zod schema annotated `ZodType<LogEntry>`), so a one-sided edit is a compile error at that schema rather than a confusing assignment failure inside `serveHostMap`. The zod schema is a module const, not a function of a domain schema — and not exported.
+- **Evidence is a fixed structural type the package owns, not a third generic.** Unlike `Failure` and `Conn`, evidence is not domain volatility: any transport that can fail can produce provenance-tagged output lines, and `@kolu/surface-remote`'s `SessionState.log` is structurally exactly this — so the producer passes its retained tail straight through, with no injection and no second evidence pipe. There is only one vocabulary to keep straight, and only one definition of it: `@kolu/surface-remote`'s `LogEntry` **is** `EvidenceLine` and its `LogEntrySchema` **is** `EvidenceLineSchema`, so "the two drifted" has no spelling rather than a pin that watches for it. Both live in the zod-only `/evidence` leaf, so a browser-bundle-constrained consumer can hold the schema without pulling the contract builder. The schema is a module const, not a function of a domain schema.
 - **Evidence rides the failure record, and the `failed` arm has no `connection` field.** `floorOnLiveness` drops `connection` over a dead link while keeping the failure record, so evidence carried there survives the floor **by construction** with the reason it belongs to, and `entry.connection?.log` on a failed entry no longer compiles. `[]` is a real value with one meaning — the failure genuinely produced no output — minted only by the seam that knows; it is never a fallback for "we couldn't see it". [Why the pairing is a shape rather than a convention](/surface/entry-contracts/) is the explanation page's job.
 - **The framework fixes only the shape, not the fields.** `failure` is whatever the map's `failure` schema (`ZodType<Failure>`) accepts — a bare string, a code-only object, or a discriminated union are all valid; the package fixes only that `failure` is a sibling field on the `failed` arm (never a second `kind`), validated by that schema. The `reason` field is padi's own convention (a human string, never parsed for control flow), not a framework guarantee — a domain whose schema carries no `reason` is equally well-formed.
 - **`EntryState<Failure, Conn>` = `EntryStatus<Failure, Conn> | { kind: "not-a-member" }`** lives in the solid-free contract module, so a node consumer re-exports it type-only.
